+++ massiv-io/src/Data/Massiv/Array/IO/Image.hs
@@ -14,17 +14,18 @@
 module Data.Massiv.Array.IO.Image
   ( Encode
   , encodeImage
+  , encodeImageM
   , imageWriteFormats
   , imageWriteAutoFormats
   , Decode
   , decodeImage
+  , decodeImageM
   , imageReadFormats
   , imageReadAutoFormats
   , module Data.Massiv.Array.IO.Image.JuicyPixels
   , module Data.Massiv.Array.IO.Image.Netpbm
   ) where
 
-import Control.Exception (throw)
 import qualified Data.ByteString as B (ByteString)
 import qualified Data.ByteString.Lazy as BL (ByteString)
 import Data.Char (toLower)
@@ -50,22 +51,48 @@ instance FileFormat (Encode (Image r cs e)) where
   exts (EncodeAs f) = exts f
 
 instance Writable (Encode (Image r cs e)) (Image r cs e) where
-  encode (EncodeAs f) _ = encode f (defaultWriteOptions f)
+  encodeM (EncodeAs f) _ = encodeM f (defaultWriteOptions f)
+
+
+-- | Encode an array into a `BL.ByteString`.
+encodeImage ::
+     (ColorSpace cs e, Source r Ix2 (Pixel cs e))
+  => [Encode (Image r cs e)]
+  -> FilePath
+  -> Image r cs e
+  -> BL.ByteString
+encodeImage formats path = either throw id . encodeImageM formats path
+{-# DEPRECATED encodeImage "In favor of a better `encodeImageM`" #-}
+
+-- | Decode a `B.ByteString` into an Array.
+decodeImage ::
+     (ColorSpace cs e, Source r Ix2 (Pixel cs e))
+  => [Decode (Image r cs e)]
+  -> FilePath
+  -> B.ByteString
+  -> Image r cs e
+decodeImage formats path = either throw fst . decodeImageM formats path
+{-# DEPRECATED decodeImage "In favor of a better `decodeImageM`" #-}
+
 
 -- | Encode an image into a lazy `BL.ByteString`, while selecting the appropriate format from the
 -- file extension.
-encodeImage
-  :: (Source r Ix2 (Pixel cs e), ColorSpace cs e)
+--
+-- @since 0.2.0
+encodeImageM
+  :: (Source r Ix2 (Pixel cs e), ColorSpace cs e, MonadThrow m)
   => [Encode (Image r cs e)] -- ^ List of image formats to choose from (useful lists are
                              -- `imageWriteFormats` and `imageWriteAutoFormats`
   -> FilePath -- ^ File name with extension, so the format can be inferred
   -> Image r cs e -- ^ Image to encode
-  -> BL.ByteString
-encodeImage formats path img = do
+  -> m BL.ByteString
+encodeImageM formats path img = do
   let ext' = P.map toLower . takeExtension $ path
   case dropWhile (not . isFormat ext') formats of
-    []    -> throw $ EncodeError $ "File format is not supported: " ++ ext'
-    (f:_) -> encode f () img
+    []    -> throwM $ EncodeError $ "File format is not supported: " ++ ext'
+    (f:_) -> encodeM f () img
+
+
 
 -- | List of image formats that can be encoded without any color space conversion.
 imageWriteFormats :: (Source r Ix2 (Pixel cs e), ColorSpace cs e) => [Encode (Image r cs e)]
@@ -113,21 +140,23 @@ instance FileFormat (Decode (Image r cs e)) where
   exts (DecodeAs f) = exts f
 
 instance Readable (Decode (Image r cs e)) (Image r cs e) where
-  decode (DecodeAs f) _ = decode f (defaultReadOptions f)
+  decodeM (DecodeAs f) _ = decodeM f (defaultReadOptions f)
 
 -- | Decode an image from the strict `ByteString` while inferring format the image is encoded in
 -- from the file extension
-decodeImage
-  :: (Source r Ix2 (Pixel cs e), ColorSpace cs e)
+--
+-- @since 0.2.0
+decodeImageM
+  :: (Source r Ix2 (Pixel cs e), ColorSpace cs e, MonadThrow m)
   => [Decode (Image r cs e)] -- ^ List of available formats to choose from
   -> FilePath -- ^ File name with extension, so format can be inferred
   -> B.ByteString -- ^ Encoded image
-  -> Image r cs e
-decodeImage formats path bs = do
+  -> m (Image r cs e, Maybe B.ByteString)
+decodeImageM formats path bs = do
   let ext' = P.map toLower . takeExtension $ path
   case dropWhile (not . isFormat ext') formats of
-    []    -> throw $ DecodeError $ "File format is not supported: " ++ ext'
-    (f:_) -> decode f () bs
+    []    -> throwM $ DecodeError $ "File format is not supported: " ++ ext'
+    (f:_) -> decodeM f () bs
 
 -- | List of image formats decodable with no colorspace conversion
 imageReadFormats