+++ massiv-io/src/Data/Massiv/Array/IO/Image/JuicyPixels.hs
@@ -78,9 +78,7 @@ import qualified Codec.Picture.ColorQuant as JP
 import qualified Codec.Picture.Gif as JP
 import qualified Codec.Picture.Jpg as JP
 import qualified Codec.Picture.Types as TypesJP
-import Control.Exception
 import Control.Monad (guard, msum, unless)
-import Data.Bifunctor
 import qualified Data.ByteString as B
 import qualified Data.ByteString.Lazy as BL (ByteString)
 import Data.Default.Class (Default(..))
@@ -121,20 +119,20 @@ instance FileFormat BMP where
 
 instance (ColorSpace cs e, Source r Ix2 (Pixel cs e)) =>
          Writable BMP (Image r cs e) where
-  encode f _ img = fromMaybeEncode f (toProxy img) $ encodeBMP img
+  encodeM f _ img = fromMaybeEncode f (toProxy img) $ encodeBMP img
 
 instance (ColorSpace cs e, ToRGBA cs e, Source r Ix2 (Pixel cs e)) =>
          Writable (Auto BMP) (Image r cs e) where
-  encode f _ = encodeAuto f encodeBMP id toPixelRGBA toPixelRGB toPixelRGBA
+  encodeM f _ = encodeAuto f encodeBMP id toPixelRGBA toPixelRGB toPixelRGBA
 
 
 instance ColorSpace cs e => Readable BMP (Image S cs e) where
-  decode f _ = fromEitherDecode f showJP fromDynamicImage . JP.decodeBitmap
+  decodeM f _ = fromEitherDecode_ f showJP fromDynamicImage . JP.decodeBitmap
 
 
 instance (Mutable r Ix2 (Pixel cs e), ColorSpace cs e) =>
          Readable (Auto BMP) (Image r cs e) where
-  decode f _ = fromEitherDecode f showJP fromAnyDynamicImage . JP.decodeBitmap
+  decodeM f _ = fromEitherDecode_ f showJP fromAnyDynamicImage . JP.decodeBitmap
 
 
 encodeBMP :: forall r cs e . (ColorSpace cs e, Source r Ix2 (Pixel cs e))
@@ -144,20 +142,20 @@ encodeBMP img =
     [ do Refl <- eqT :: Maybe (cs :~: Y)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeBitmap $ toJPImageY8 img
-           , return $ JP.encodeBitmap $ toJPImageY8 $ M.map toWord8 img
+                pure $ JP.encodeBitmap $ toJPImageY8 img
+           , pure $ JP.encodeBitmap $ toJPImageY8 $ M.map toWord8 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: RGB)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeBitmap $ toJPImageRGB8 img
-           , return $ JP.encodeBitmap $ toJPImageRGB8 $ M.map toWord8 img
+                pure $ JP.encodeBitmap $ toJPImageRGB8 img
+           , pure $ JP.encodeBitmap $ toJPImageRGB8 $ M.map toWord8 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: RGBA)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeBitmap $ toJPImageRGBA8 img
-           , return $ JP.encodeBitmap $ toJPImageRGBA8 $ M.map toWord8 img
+                pure $ JP.encodeBitmap $ toJPImageRGBA8 img
+           , pure $ JP.encodeBitmap $ toJPImageRGBA8 $ M.map toWord8 img
            ]
     ]
 
@@ -175,20 +173,20 @@ instance FileFormat PNG where
 
 instance (ColorSpace cs e, Source r Ix2 (Pixel cs e)) =>
          Writable PNG (Image r cs e) where
-  encode f _ img = fromMaybeEncode f (toProxy img) (encodePNG img)
+  encodeM f _ img = fromMaybeEncode f (toProxy img) (encodePNG img)
 
 
 instance (ColorSpace cs e, ToYA cs e, ToRGBA cs e, Source r Ix2 (Pixel cs e)) =>
          Writable (Auto PNG) (Image r cs e) where
-  encode f _ = encodeAuto f encodePNG id toPixelYA toPixelRGB toPixelRGBA
+  encodeM f _ = encodeAuto f encodePNG id toPixelYA toPixelRGB toPixelRGBA
 
 
 instance ColorSpace cs e => Readable PNG (Image S cs e) where
-  decode f _ = fromEitherDecode f showJP fromDynamicImage . JP.decodePng
+  decodeM f _ = fromEitherDecode_ f showJP fromDynamicImage . JP.decodePng
 
 instance (Mutable r Ix2 (Pixel cs e), ColorSpace cs e) =>
          Readable (Auto PNG) (Image r cs e) where
-  decode f _ = fromEitherDecode f showJP fromAnyDynamicImage . JP.decodePng
+  decodeM f _ = fromEitherDecode_ f showJP fromAnyDynamicImage . JP.decodePng
 
 
 
@@ -199,34 +197,34 @@ encodePNG img =
     [ do Refl <- eqT :: Maybe (cs :~: Y)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodePng $ toJPImageY8 img
+                pure $ JP.encodePng $ toJPImageY8 img
            , do Refl <- eqT :: Maybe (e :~: Word16)
-                return $ JP.encodePng $ toJPImageY16 img
-           , return $ JP.encodePng $ toJPImageY16 $ M.map toWord16 img
+                pure $ JP.encodePng $ toJPImageY16 img
+           , pure $ JP.encodePng $ toJPImageY16 $ M.map toWord16 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: YA)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodePng $ toJPImageYA8 img
+                pure $ JP.encodePng $ toJPImageYA8 img
            , do Refl <- eqT :: Maybe (e :~: Word16)
-                return $ JP.encodePng $ toJPImageYA16 img
-           , return $ JP.encodePng $ toJPImageYA16 $ M.map toWord16 img
+                pure $ JP.encodePng $ toJPImageYA16 img
+           , pure $ JP.encodePng $ toJPImageYA16 $ M.map toWord16 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: RGB)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodePng $ toJPImageRGB8 img
+                pure $ JP.encodePng $ toJPImageRGB8 img
            , do Refl <- eqT :: Maybe (e :~: Word16)
-                return $ JP.encodePng $ toJPImageRGB16 img
-           , return $ JP.encodePng $ toJPImageRGB16 $ M.map toWord16 img
+                pure $ JP.encodePng $ toJPImageRGB16 img
+           , pure $ JP.encodePng $ toJPImageRGB16 $ M.map toWord16 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: RGBA)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodePng $ toJPImageRGBA8 img
+                pure $ JP.encodePng $ toJPImageRGBA8 img
            , do Refl <- eqT :: Maybe (e :~: Word16)
-                return $ JP.encodePng $ toJPImageRGBA16 img
-           , return $ JP.encodePng $ toJPImageRGBA16 $ M.map toWord16 img
+                pure $ JP.encodePng $ toJPImageRGBA16 img
+           , pure $ JP.encodePng $ toJPImageRGBA16 $ M.map toWord16 img
            ]
     ]
 
@@ -285,93 +283,97 @@ instance FileFormat (Sequence (Auto GIF)) where
 
 instance (ColorSpace cs e, Source r Ix2 (Pixel cs e)) =>
          Writable GIF (Image r cs e) where
-  encode f opt img = fromMaybeEncode f (toProxy img) $ encodeGIF opt img
+  encodeM f opt img =
+    encodeError =<< fromMaybeEncode f (toProxy img) (encodeGIF opt img)
 
 instance (ColorSpace cs e, ToY cs e, ToRGB cs e, Source r Ix2 (Pixel cs e)) =>
          Writable (Auto GIF) (Image r cs e) where
-  encode f opt = encodeAuto f (encodeGIF opt) id toPixelY toPixelRGB toPixelRGB
+  encodeM f opt img =
+    encodeError =<< encodeAuto f (encodeGIF opt) id toPixelY toPixelRGB toPixelRGB img
 
 
 instance ColorSpace cs e => Readable GIF (Image S cs e) where
-  decode f _ = fromEitherDecode f showJP fromDynamicImage . JP.decodeGif
+  decodeM f _ = fromEitherDecode_ f showJP fromDynamicImage . JP.decodeGif
 
 instance (Mutable r Ix2 (Pixel cs e), ColorSpace cs e) =>
          Readable (Auto GIF) (Image r cs e) where
-  decode f _ = fromEitherDecode f showJP fromAnyDynamicImage . JP.decodeGif
-
-
+  decodeM f _ = fromEitherDecode_ f showJP fromAnyDynamicImage . JP.decodeGif
 
 
 instance ColorSpace cs e =>
          Readable (Sequence GIF) (Array B Ix1 (Image S cs e)) where
-  decode f _ = decodeGIFs f fromDynamicImage
+  decodeM f _ = decodeGIFs f fromDynamicImage
 
 instance (Mutable r Ix2 (Pixel cs e), ColorSpace cs e) =>
          Readable (Sequence (Auto GIF)) (Array B Ix1 (Image r cs e)) where
-  decode f _ = decodeGIFs f fromAnyDynamicImage
+  decodeM f _ = decodeGIFs f fromAnyDynamicImage
 
 
 instance (ColorSpace cs e, Source r Ix2 (Pixel cs e)) =>
          Writable (Sequence GIF) (Array B Ix1 (JP.GifDelay, Image r cs e)) where
-  encode _ opts =
-    fromMaybeEncode (Sequence GIF) (Proxy :: Proxy (Image r cs e)) . encodeGIFs opts
+  encodeM f opts img = do
+    encodeError =<< fromMaybeEncode f (Proxy :: Proxy (Image r cs e)) (encodeGIFs opts img)
 
 
 instance ColorSpace cs e =>
          Readable (Sequence GIF) (Array B Ix1 (JP.GifDelay, Image S cs e)) where
-  decode f _ = decodeGIFsWithDelays f fromDynamicImage
+  decodeM f _ = decodeGIFsWithDelays f fromDynamicImage
 
 
+fromEitherDecode_ ::
+     (MonadThrow m, ColorSpace cs e, FileFormat f, Typeable r)
+  => f
+  -> (a -> String)
+  -> (a -> Maybe (Image r cs e))
+  -> Either String a
+  -> m (Image r cs e, Maybe B.ByteString)
+fromEitherDecode_ f showCS conv eImg = do
+  jImg <- either (throwM . DecodeError) pure eImg
+  img <- fromEitherDecode f showCS conv jImg
+  pure (img, Nothing)
 
 -- Animated GIF Format frames reading into an Array of Images
 
 decodeGIFs
-  :: (FileFormat f, Mutable r Ix2 (Pixel cs e), ColorSpace cs e)
+  :: (FileFormat f, Mutable r Ix2 (Pixel cs e), ColorSpace cs e, MonadThrow m)
   => f
   -> (JP.DynamicImage -> Maybe (Image r cs e))
   -> B.ByteString
-  -> Array B Ix1 (Image r cs e)
-decodeGIFs f converter bs =
-  either throw (fromList Seq) $ do
-    jpImgs <- first (toException . DecodeError) $ JP.decodeGifImages bs
-    first toException $ P.mapM (convertEither f showJP converter) jpImgs
-  -- either
-  --   (throw . DecodeError)
-  --   (fromList' Seq)
-  --   (P.map (fromEitherDecode f showJP decoder . Right) <$> JP.decodeGifImages bs)
+  -> m (Array B Ix1 (Image r cs e), Maybe B.ByteString)
+decodeGIFs f converter bs = do
+  jpImgs <- either (throwM . DecodeError) pure $ JP.decodeGifImages bs
+  imgs <- P.traverse (convertEither f showJP converter) jpImgs
+  pure (fromList Seq imgs, Nothing)
 {-# INLINE decodeGIFs #-}
 
 
-
 decodeGIFsWithDelays
-  :: ColorSpace cs e
+  :: (ColorSpace cs e, MonadThrow m)
   => Sequence GIF
   -> (JP.DynamicImage -> Maybe (Image S cs e))
   -> B.ByteString
-  -> Array B Ix1 (JP.GifDelay, Image S cs e)
-decodeGIFsWithDelays f converter bs =
-  either throw (fromList Seq) $ do
-    jpImgsLs <- first (toException . DecodeError) $ JP.decodeGifImages bs
-    delays <- first (toException . DecodeError) $ JP.getDelaysGifImages bs
-    imgs <- first toException $ P.mapM (convertEither f showJP converter) jpImgsLs
-    return $ P.zip delays imgs
+  -> m (Array B Ix1 (JP.GifDelay, Image S cs e), Maybe B.ByteString)
+decodeGIFsWithDelays f converter bs = do
+  jpImgsLs <- either (throwM . DecodeError) pure $ JP.decodeGifImages bs
+  delays <- either (throwM . DecodeError) pure $ JP.getDelaysGifImages bs
+  imgs <- P.traverse (convertEither f showJP converter) jpImgsLs
+  pure (fromList Seq (P.zip delays imgs), Nothing)
 {-# INLINE decodeGIFsWithDelays #-}
-
 
 
 encodeGIF :: forall r cs e . (ColorSpace cs e, Source r Ix2 (Pixel cs e))
           => WriteOptionsGIF
           -> Image r cs e
-          -> Maybe BL.ByteString
+          -> Maybe (Either String BL.ByteString)
 encodeGIF (WriteOptionsGIF pal) img =
-  (either (throw . EncodeError) id . uncurry JP.encodeGifImageWithPalette) <$>
+  uncurry JP.encodeGifImageWithPalette <$>
   msum
     [ do Refl <- eqT :: Maybe (cs :~: Y)
          jImg <-
            msum
              [ do Refl <- eqT :: Maybe (e :~: Word8)
-                  return $ toJPImageY8 img
-             , return $ toJPImageY8 $ M.map toWord8 img
+                  pure $ toJPImageY8 img
+             , pure $ toJPImageY8 $ M.map toWord8 img
              ]
          return (jImg, JP.greyPalette)
     , do Refl <- eqT :: Maybe (cs :~: RGB)
@@ -386,15 +388,15 @@ encodeGIF (WriteOptionsGIF pal) img =
 encodeGIFs :: forall r cs e . (ColorSpace cs e, Source r Ix2 (Pixel cs e))
            => WriteOptionsSequenceGIF
            -> Array B Ix1 (JP.GifDelay, Image r cs e)
-           -> Maybe BL.ByteString
+           -> Maybe (Either String BL.ByteString)
 encodeGIFs (WriteOptionsSequenceGIF pal looping) arr = do
   palImgsLs <-
     msum
       [ do Refl <- eqT :: Maybe (cs :~: Y)
            msum
              [ do Refl <- eqT :: Maybe (e :~: Word8)
-                  return $ P.map ((, JP.greyPalette) . toJPImageY8) imgsLs
-             , return $
+                  pure $ P.map ((, JP.greyPalette) . toJPImageY8) imgsLs
+             , pure $
                P.map ((, JP.greyPalette) . toJPImageY8 . M.map toWord8) imgsLs
              ]
       , do Refl <- eqT :: Maybe (cs :~: RGB)
@@ -405,8 +407,7 @@ encodeGIFs (WriteOptionsSequenceGIF pal looping) arr = do
              ]
       ]
   let palDelImgsLs = P.zipWith (\(i, p) d -> (p, d, i)) palImgsLs delaysLs
-  return $
-    either (throw . EncodeError) id $ JP.encodeGifImages looping palDelImgsLs
+  pure $ JP.encodeGifImages looping palDelImgsLs
   where
     delaysLs = toList delays
     imgsLs = toList imgs
@@ -421,8 +422,8 @@ palettizeRGB :: forall r e . (ColorSpace RGB e, Source r Ix2 (Pixel RGB e))
 palettizeRGB pal img =
   msum
     [ do Refl <- eqT :: Maybe (e :~: Word8)
-         return $ palettize' img
-    , return $ palettize' $ M.map toWord8 img
+         pure $ palettize' img
+    , pure $ palettize' $ M.map toWord8 img
     ]
   where
     palettize' :: forall r' . Source r' Ix2 (Pixel RGB Word8) =>
@@ -448,21 +449,21 @@ instance FileFormat HDR where
 
 instance (ColorSpace cs e, Source r Ix2 (Pixel cs e)) =>
          Writable HDR (Image r cs e) where
-  encode f _ img = fromMaybeEncode f (toProxy img) $ encodeHDR img
+  encodeM f _ img = fromMaybeEncode f (toProxy img) $ encodeHDR img
 
 
 instance (ColorSpace cs e, ToRGB cs e, Source r Ix2 (Pixel cs e)) =>
          Writable (Auto HDR) (Image r cs e) where
-  encode f _ =
+  encodeM f _ =
     encodeAuto f encodeHDR toPixelRGB toPixelRGB toPixelRGB toPixelRGB
 
 
 instance ColorSpace cs e => Readable HDR (Image S cs e) where
-  decode f _ = fromEitherDecode f showJP fromDynamicImage . JP.decodeHDR
+  decodeM f _ = fromEitherDecode_ f showJP fromDynamicImage . JP.decodeHDR
 
 instance (Mutable r Ix2 (Pixel cs e), ColorSpace cs e) =>
          Readable (Auto HDR) (Image r cs e) where
-  decode f _ = fromEitherDecode f showJP fromAnyDynamicImage . JP.decodeHDR
+  decodeM f _ = fromEitherDecode_ f showJP fromAnyDynamicImage . JP.decodeHDR
 
 
 
@@ -473,8 +474,8 @@ encodeHDR img = do
   Refl <- eqT :: Maybe (cs :~: RGB)
   msum
     [ do Refl <- eqT :: Maybe (e :~: Float)
-         return $ JP.encodeHDR $ toJPImageRGBF img
-    , return $ JP.encodeHDR $ toJPImageRGBF $ M.map toFloat img
+         pure $ JP.encodeHDR $ toJPImageRGBF img
+    , pure $ JP.encodeHDR $ toJPImageRGBF $ M.map toFloat img
     ]
 
 
@@ -506,21 +507,21 @@ instance FileFormat JPG where
 
 instance (ColorSpace cs e, Source r Ix2 (Pixel cs e)) =>
          Writable JPG (Image r cs e) where
-  encode f opts img = fromMaybeEncode f (toProxy img) $ encodeJPG opts img
+  encodeM f opts img = fromMaybeEncode f (toProxy img) $ encodeJPG opts img
 
 
 instance (ColorSpace cs e, ToYCbCr cs e, Source r Ix2 (Pixel cs e)) =>
          Writable (Auto JPG) (Image r cs e) where
-  encode f opt =
+  encodeM f opt =
     encodeAuto f (encodeJPG opt) toPixelYCbCr toPixelYCbCr toPixelYCbCr toPixelYCbCr
 
 
 instance ColorSpace cs e => Readable JPG (Image S cs e) where
-  decode f _ = fromEitherDecode f showJP fromDynamicImage . JP.decodeJpeg
+  decodeM f _ = fromEitherDecode_ f showJP fromDynamicImage . JP.decodeJpeg
 
 instance (Mutable r Ix2 (Pixel cs e), ColorSpace cs e) =>
          Readable (Auto JPG) (Image r cs e) where
-  decode f _ = fromEitherDecode f showJP fromAnyDynamicImage . JP.decodeJpeg
+  decodeM f _ = fromEitherDecode_ f showJP fromAnyDynamicImage . JP.decodeJpeg
 
 
 
@@ -532,26 +533,26 @@ encodeJPG (WriteOptionsJPG q) img =
     [ do Refl <- eqT :: Maybe (cs :~: Y)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ encJPG $ toJPImageY8 img
-           , return $ encJPG $ toJPImageY8 $ M.map toWord8 img
+                pure $ encJPG $ toJPImageY8 img
+           , pure $ encJPG $ toJPImageY8 $ M.map toWord8 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: RGB)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ encJPG $ toJPImageRGB8 img
-           , return $ encJPG $ toJPImageRGB8 $ M.map toWord8 img
+                pure $ encJPG $ toJPImageRGB8 img
+           , pure $ encJPG $ toJPImageRGB8 $ M.map toWord8 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: CMYK)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ encJPG $ toJPImageCMYK8 img
-           , return $ encJPG $ toJPImageCMYK8 $ M.map toWord8 img
+                pure $ encJPG $ toJPImageCMYK8 img
+           , pure $ encJPG $ toJPImageCMYK8 $ M.map toWord8 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: YCbCr)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ encJPG $ toJPImageYCbCr8 img
-           , return $ encJPG $ toJPImageYCbCr8 $ M.map toWord8 img
+                pure $ encJPG $ toJPImageYCbCr8 img
+           , pure $ encJPG $ toJPImageYCbCr8 $ M.map toWord8 img
            ]
     ]
   where
@@ -578,20 +579,20 @@ instance FileFormat TGA where
 
 instance (ColorSpace cs e, Source r Ix2 (Pixel cs e)) =>
          Writable TGA (Image r cs e) where
-  encode f _ img = fromMaybeEncode f (toProxy img) $ encodeTGA img
+  encodeM f _ img = fromMaybeEncode f (toProxy img) $ encodeTGA img
 
 instance (ColorSpace cs e, ToRGBA cs e, Source r Ix2 (Pixel cs e)) =>
          Writable (Auto TGA) (Image r cs e) where
-  encode f _ = encodeAuto f encodeTGA id toPixelRGBA toPixelRGB toPixelRGBA
+  encodeM f _ = encodeAuto f encodeTGA id toPixelRGBA toPixelRGB toPixelRGBA
 
 
 instance ColorSpace cs e => Readable TGA (Image S cs e) where
-  decode f _ = fromEitherDecode f showJP fromDynamicImage . JP.decodeTga
+  decodeM f _ = fromEitherDecode_ f showJP fromDynamicImage . JP.decodeTga
 
 
 instance (Mutable r Ix2 (Pixel cs e), ColorSpace cs e) =>
          Readable (Auto TGA) (Image r cs e) where
-  decode f _ = fromEitherDecode f showJP fromAnyDynamicImage . JP.decodeTga
+  decodeM f _ = fromEitherDecode_ f showJP fromAnyDynamicImage . JP.decodeTga
 
 
 encodeTGA :: forall r cs e . (ColorSpace cs e, Source r Ix2 (Pixel cs e))
@@ -601,20 +602,20 @@ encodeTGA img =
     [ do Refl <- eqT :: Maybe (cs :~: Y)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeTga $ toJPImageY8 img
-           , return $ JP.encodeTga $ toJPImageY8 $ M.map toWord8 img
+                pure $ JP.encodeTga $ toJPImageY8 img
+           , pure $ JP.encodeTga $ toJPImageY8 $ M.map toWord8 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: RGB)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeTga $ toJPImageRGB8 img
-           , return $ JP.encodeTga $ toJPImageRGB8 $ M.map toWord8 img
+                pure $ JP.encodeTga $ toJPImageRGB8 img
+           , pure $ JP.encodeTga $ toJPImageRGB8 $ M.map toWord8 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: RGBA)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeTga $ toJPImageRGBA8 img
-           , return $ JP.encodeTga $ toJPImageRGBA8 $ M.map toWord8 img
+                pure $ JP.encodeTga $ toJPImageRGBA8 img
+           , pure $ JP.encodeTga $ toJPImageRGBA8 $ M.map toWord8 img
            ]
     ]
 
@@ -638,20 +639,20 @@ instance FileFormat TIF where
 
 instance (ColorSpace cs e, Source r Ix2 (Pixel cs e)) =>
          Writable TIF (Image r cs e) where
-  encode f _ img = fromMaybeEncode f (toProxy img) $ encodeTIF img
+  encodeM f _ img = fromMaybeEncode f (toProxy img) $ encodeTIF img
 
 
 instance (ColorSpace cs e, ToRGBA cs e, Source r Ix2 (Pixel cs e)) =>
          Writable (Auto TIF) (Image r cs e) where
-  encode f _ = encodeAuto f encodeTIF id id id toPixelRGBA
+  encodeM f _ = encodeAuto f encodeTIF id id id toPixelRGBA
 
 
 instance ColorSpace cs e => Readable TIF (Image S cs e) where
-  decode f _ = fromEitherDecode f showJP fromDynamicImage . JP.decodeTiff
+  decodeM f _ = fromEitherDecode_ f showJP fromDynamicImage . JP.decodeTiff
 
 instance (Mutable r Ix2 (Pixel cs e), ColorSpace cs e) =>
          Readable (Auto TIF) (Image r cs e) where
-  decode f _ = fromEitherDecode f showJP fromAnyDynamicImage . JP.decodeTiff
+  decodeM f _ = fromEitherDecode_ f showJP fromAnyDynamicImage . JP.decodeTiff
 
 
 
@@ -663,56 +664,56 @@ encodeTIF img =
     [ do Refl <- eqT :: Maybe (cs :~: Y)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeTiff $ toJPImageY8 img
+                pure $ JP.encodeTiff $ toJPImageY8 img
            , do Refl <- eqT :: Maybe (e :~: Word16)
-                return $ JP.encodeTiff $ toJPImageY16 img
+                pure $ JP.encodeTiff $ toJPImageY16 img
 #if MIN_VERSION_JuicyPixels(3,3,0)
            , do Refl <- eqT :: Maybe (e :~: Word32)
-                return $ JP.encodeTiff $ toJPImageY32 img
+                pure $ JP.encodeTiff $ toJPImageY32 img
            , do Refl <- eqT :: Maybe (e :~: Float)
-                return $ JP.encodeTiff $ toJPImageYF img
+                pure $ JP.encodeTiff $ toJPImageYF img
 #endif
-           , return $ JP.encodeTiff $ toJPImageY16 $ M.map toWord16 img
+           , pure $ JP.encodeTiff $ toJPImageY16 $ M.map toWord16 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: YA)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeTiff $ toJPImageYA8 img
+                pure $ JP.encodeTiff $ toJPImageYA8 img
            , do Refl <- eqT :: Maybe (e :~: Word16)
-                return $ JP.encodeTiff $ toJPImageYA16 img
-           , return $ JP.encodeTiff $ toJPImageYA16 $ M.map toWord16 img
+                pure $ JP.encodeTiff $ toJPImageYA16 img
+           , pure $ JP.encodeTiff $ toJPImageYA16 $ M.map toWord16 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: RGB)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeTiff $ toJPImageRGB8 img
+                pure $ JP.encodeTiff $ toJPImageRGB8 img
            , do Refl <- eqT :: Maybe (e :~: Word16)
-                return $ JP.encodeTiff $ toJPImageRGB16 img
-           , return $ JP.encodeTiff $ toJPImageRGB16 $ M.map toWord16 img
+                pure $ JP.encodeTiff $ toJPImageRGB16 img
+           , pure $ JP.encodeTiff $ toJPImageRGB16 $ M.map toWord16 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: RGBA)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeTiff $ toJPImageRGBA8 img
+                pure $ JP.encodeTiff $ toJPImageRGBA8 img
            , do Refl <- eqT :: Maybe (e :~: Word16)
-                return $ JP.encodeTiff $ toJPImageRGBA16 img
-           , return $ JP.encodeTiff $ toJPImageRGBA16 $ M.map toWord16 img
+                pure $ JP.encodeTiff $ toJPImageRGBA16 img
+           , pure $ JP.encodeTiff $ toJPImageRGBA16 $ M.map toWord16 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: YCbCr)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeTiff $ toJPImageYCbCr8 img
-           , return $ JP.encodeTiff $ toJPImageYCbCr8 $ M.map toWord8 img
+                pure $ JP.encodeTiff $ toJPImageYCbCr8 img
+           , pure $ JP.encodeTiff $ toJPImageYCbCr8 $ M.map toWord8 img
            ]
     , do Refl <- eqT :: Maybe (cs :~: CMYK)
          msum
            [ do Refl <- eqT :: Maybe (e :~: Word8)
-                return $ JP.encodeTiff $ toJPImageCMYK8 img
+                pure $ JP.encodeTiff $ toJPImageCMYK8 img
            , do Refl <- eqT :: Maybe (e :~: Word16)
-                return $ JP.encodeTiff $ toJPImageCMYK16 img
+                pure $ JP.encodeTiff $ toJPImageCMYK16 img
              -- for CMYK default is 8bit, instead of 16bit, since many viewers and editors
              -- don't support it.
-           , return $ JP.encodeTiff $ toJPImageCMYK8 $ M.map toWord8 img
+           , pure $ JP.encodeTiff $ toJPImageCMYK8 $ M.map toWord8 img
            ]
     ]
 
@@ -723,77 +724,6 @@ encodeTIF img =
 -- Common encoding/decoding functions ------------------------------------------
 --------------------------------------------------------------------------------
 
-
-encodeAuto
-  :: forall f r cs e a csY eY csYA eYA csC eC csCA eCA.
-     ( ColorSpace cs e
-     , ColorSpace csC eC
-     , ColorSpace csCA eCA
-     , ColorSpace csY eY
-     , ColorSpace csYA eYA
-     , Source r Ix2 (Pixel cs e)
-     , FileFormat f
-     )
-  => f
-  -> (forall r' cs' e'. (Source r' Ix2 (Pixel cs' e'), ColorSpace cs' e') =>
-                          Image r' cs' e' -> Maybe a)
-  -> (Pixel cs e -> Pixel csY eY) -- ^ To preferred from Luma
-  -> (Pixel cs e -> Pixel csYA eYA) -- ^ To preferred from Luma with Alpha
-  -> (Pixel cs e -> Pixel csC eC) -- ^ To preferred from any color
-  -> (Pixel cs e -> Pixel csCA eCA) -- ^ To preferred from any color with Alpha
-  -> Image r cs e
-  -> a
-encodeAuto f enc toLuma toLumaA toColor toColorA img =
-  fromMaybeEncode f (toProxy img) $ msum
-    [ enc img
-    , do Refl <- eqT :: Maybe (cs :~: Y)
-         enc $ M.map toLuma img
-    , do Refl <- eqT :: Maybe (cs :~: YA)
-         enc $ M.map toLumaA img
-    , do Refl <- eqT :: Maybe (cs :~: RGB)
-         enc $ M.map toColor img
-    , do Refl <- eqT :: Maybe (cs :~: RGBA)
-         enc $ M.map toColorA img
-    , do Refl <- eqT :: Maybe (cs :~: HSI)
-         enc $ M.map toColor img
-    , do Refl <- eqT :: Maybe (cs :~: HSIA)
-         enc $ M.map toColorA img
-    , do Refl <- eqT :: Maybe (cs :~: YCbCr)
-         enc $ M.map toColor img
-    , do Refl <- eqT :: Maybe (cs :~: YCbCrA)
-         enc $ M.map toColorA img
-    , do Refl <- eqT :: Maybe (cs :~: CMYK)
-         enc $ M.map toColor img
-    , do Refl <- eqT :: Maybe (cs :~: CMYKA)
-         enc $ M.map toColorA img
-    , do Refl <- eqT :: Maybe (Pixel cs e :~: Pixel X Bit)
-         enc $ M.map fromPixelBinary img
-    ]
-
-
-
-
-elevate
-  :: forall cs e' e.
-     ( ColorSpace cs e'
-     , ColorSpace cs e
-     , Source D Ix2 (Pixel cs e')
-     )
-  => Image D cs e' -> Maybe (Image D cs e)
-elevate img =
-  msum
-    [ fmap (\Refl -> img) (eqT :: Maybe (e :~: e'))
-    , do Refl <- eqT :: Maybe (e :~: Word8)
-         return $ M.map toWord8 img
-    , do Refl <- eqT :: Maybe (e :~: Word16)
-         return $ M.map toWord16 img
-    , do Refl <- eqT :: Maybe (e :~: Word32)
-         return $ M.map toWord32 img
-    , do Refl <- eqT :: Maybe (e :~: Word64)
-         return $ M.map toWord64 img
-    , do Refl <- eqT :: Maybe (e :~: Double)
-         return $ M.map toDouble img
-    ]
 
 fromDynamicImage :: forall cs e . ColorSpace cs e
                  => JP.DynamicImage -> Maybe (Image S cs e)
@@ -868,47 +798,6 @@ fromAnyDynamicImage jpDynImg =
     JP.ImageCMYK16 jimg -> (fromJPImageUnsafe jimg :: Maybe (Image S CMYK Word16)) >>= toAnyCS
 
 
-toAnyCS
-  :: forall r' cs' e' r cs e.
-     ( Source r' Ix2 (Pixel cs' e')
-     , Mutable r Ix2 (Pixel cs e)
-     , ColorSpace cs e
-     , ToYA cs' e'
-     , ToRGBA cs' e'
-     , ToHSIA cs' e'
-     , ToCMYKA cs' e'
-     , ToYCbCrA cs' e'
-     )
-  => Image r' cs' e' -> Maybe (Image r cs e)
-toAnyCS img =
-  msum
-    [ (\Refl -> computeSource img) <$>
-      (eqT :: Maybe (Pixel cs' e' :~: Pixel cs e))
-    , do Refl <- eqT :: Maybe (cs :~: Y)
-         compute <$> elevate (M.map toPixelY img)
-    , do Refl <- eqT :: Maybe (cs :~: YA)
-         compute <$> elevate (M.map toPixelYA img)
-    , do Refl <- eqT :: Maybe (cs :~: RGB)
-         compute <$> elevate (M.map toPixelRGB img)
-    , do Refl <- eqT :: Maybe (cs :~: RGBA)
-         compute <$> elevate (M.map toPixelRGBA img)
-    , do Refl <- eqT :: Maybe (cs :~: HSI)
-         compute <$> elevate (M.map toPixelHSI img)
-    , do Refl <- eqT :: Maybe (cs :~: HSIA)
-         compute <$> elevate (M.map toPixelHSIA img)
-    , do Refl <- eqT :: Maybe (cs :~: CMYK)
-         compute <$> elevate (M.map toPixelCMYK img)
-    , do Refl <- eqT :: Maybe (cs :~: CMYKA)
-         compute <$> elevate (M.map toPixelCMYKA img)
-    , do Refl <- eqT :: Maybe (cs :~: YCbCr)
-         compute <$> elevate (M.map toPixelYCbCr img)
-    , do Refl <- eqT :: Maybe (cs :~: YCbCrA)
-         compute <$> elevate (M.map toPixelYCbCrA img)
-    , do Refl <- eqT :: Maybe (Pixel cs e :~: Pixel X Bit)
-         return $ compute $ M.map toPixelBinary img
-    ]
-
-
 
 showJP :: JP.DynamicImage -> String
 showJP (JP.ImageY8     _) = "Image S Y Word8"
@@ -936,9 +825,10 @@ toJPImageUnsafe
                       ColorSpace cs (JP.PixelBaseComponent a))
   => Image r cs (JP.PixelBaseComponent a)
   -> JP.Image a
-toJPImageUnsafe img = JP.Image n m $ V.unsafeCast $ toVector arrS where
-  !arrS = computeSource img :: Image S cs (JP.PixelBaseComponent a)
-  Sz (m :. n) = size img
+toJPImageUnsafe img = JP.Image n m $ V.unsafeCast $ toVector arrS
+  where
+    !arrS = computeSource img :: Image S cs (JP.PixelBaseComponent a)
+    Sz (m :. n) = size img
 {-# INLINE toJPImageUnsafe #-}
 
 toJPImageY8 :: Source r Ix2 (Pixel Y Word8) => Image r Y Word8 -> JP.Image TypesJP.Pixel8