+++ massiv-io/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.2.0
+
+* Switch to `MonadIO`
+* Write files with durability and atomicity guarantees
+* Switch to `encodeM` and `decodeM`, as well as corresponding `encodeImagM` and `decodeImageM`
+* Addition of `ConvertError`, `DecodeError` and `EncodeError`.
+
+
 # 0.1.9
 
 * Fix `HDR` decoding, i.e. `.hdr` and `.pic` file reading.